simplify print_docs_diff.py
